Updated README, re-organized code, modified sounds
