+++ bin/merge.js
@@ -1,8 +1,8 @@
 var fs = require('fs'),
-    chunks = fs.readdirSync('./recordings'),
+    chunks = fs.readdirSync('../recordings'),
     inputStream,
     currentfile,
-    outputStream = fs.createWriteStream('./recordings/merge.pcm');
+    outputStream = fs.createWriteStream('../recordings/merge.pcm');
 
 chunks.sort((a, b) => { return a - b; });
 
@@ -12,14 +12,14 @@ function appendFiles() {
         return;
     }
 
-    currentfile = './recordings/' + chunks.shift();
+    currentfile = '../recordings/' + chunks.shift();
     inputStream = fs.createReadStream(currentfile);
-    
+
     inputStream.pipe(outputStream, { end: false });
 
     inputStream.on('end', function() {
         console.log(currentfile + ' appended');
-        appendFiles();        
+        appendFiles();
     });
 }
