+++ sounds/LICENSE
@@ -1,0 +1,15 @@
+
+Hello from Orange Free Sounds,
+
+
+Stock audio - Free sound effects, loops and music.
+ 
+
+There are no hidden costs or need to sign-up. 
+
+
+
+License: The sound effect is permitted for non-commercial use under license “Attribution-NonCommercial 4.0 International (CC BY-NC 4.0)
+
+
+http://www.orangefreesounds.com/
